requirements.txt
# Core Dependencies
torch>=2.0.0
pennylane>=0.33.0
pennylane-lightning>=0.33.0
numpy>=1.24.0
scipy>=1.10.0
gymnasium>=0.29.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
hydra-core>=1.3.0

# ML Experiment Tracking
mlflow>=2.8.0
tensorboard>=2.14.0

# Quantum backends (optional)
pennylane-qiskit>=0.33.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebooks
jupyter>=1.0.0
ipython>=8.12.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Stable Baselines (reference only)
stable-baselines3>=2.1.0
